Move array_from_numpy to flwr.compat
